requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
distro==1.9.0
fastapi==0.135.2
ffmpy==1.0.0
filelock==3.25.2
filetype==1.2.0
fsspec==2026.2.0
google-auth==2.49.1
google-genai==1.68.0
gradio==6.9.0
gradio-client==2.3.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.7.2
idna==3.11
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.21
langchain-google-genai==4.2.1
langsmith==0.7.22
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
numpy==2.4.3
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
pyyaml==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
tomlkit==0.13.3
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.42.0
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
